Task 1: fix #3

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -8,7 +8,7 @@
  *
  * Return: a pointer to dest
  */
-char *_memcpy(char *dest, char *src, unsigned int n)
+char *_memcpy(char *dest, char *src, size_t n)
 {
 	unsigned int i;
